fixed the init states of the RNG streams for each task
- implemented function ‘set_RNG_stream’ to set the state of the RNG on
the nodes from the prepared streams. This is necessary because RNGkind
is likely not set correctly and set.seed only handles single integer
values.
- Function behaviors:
- setting RNGkind(kind = "XXX") updates .Random.seed to match kind, but
this update is not reversible, i.e., switching from kind = X1 to X2 and
back to X1 does not get back to the same .Random.seed
- assigning an appropriate value to .Random.seed updates RNGkind()
accordingly
- setting RNGkind in any function carries forward to child-, parent-,
and the global environment
- RNGkind() on the cluster at execution time does not and is not
affected by the global environment's RNGkind()
- RNGkind() on each cluster remains to the next function call
- the initial state of RNGkind() on the cluster
    - cluster type = "FORK: then it depends on what RNGkind() was in
the environment when the cluster was created
    - cluster type = "SOCK": then it will be the default value, i.e.,
"Mersenne-Twister"

Other stuff:
- implemented function ‘dir_create’: The function returns invisible
TRUE  if the path exists
or if the path could successfully (and recursively by default) be
created. If the path does not exist and the path cannot
be created it throws an error with an useful message.